Stable2

+++ src/java/com/angstore/services/ProductService.java
@@ -64,7 +64,7 @@ public class ProductService {
         EntityManager em = emf.createEntityManager();
         
         ArrayList<Product> products = new ArrayList<>();
-        for(int i = 26; i <= 31; i++) {
+        for(long i = 26; i <= 31; i++) {
             products.add(em.find(Product.class, i));
         }
         return products;
